Introduce BikeSearch base component
