+++ app/javascript/packs/bikes_search/components/PartialSerialSearch.js
@@ -1,74 +1,14 @@
 /* eslint import/no-unresolved: 0 */
 
-import React, { Fragment, Component } from "react";
+import React from "react";
 
-import BikeSearchResult from "./BikeSearchResult";
+import BikeSearch from "./BikeSearch";
 import { fetchPartialMatchSearch } from "../../api";
-import Loading from "../../Loading";
-import honeybadger from "../../utils/honeybadger";
-import TimeParser from "../../utils/time_parser";
 
-class PartialSerialSearch extends Component {
-  state = {
-    loading: false,
-    results: []
-  }
-
-  componentWillMount() {
-    this.resultsBeingFetched();
-    fetchPartialMatchSearch(this.props.interpretedParams)
-      .then(this.resultsFetched)
-      .catch(this.handleError);
-  }
-
-  componentDidUpdate() {
-    TimeParser().localize();
-  }
-
-  resultsBeingFetched = () => {
-    this.setState({ loading: true });
-    this.toggleHeader({ isLoading: true });
-  }
-
-  resultsFetched = ({ bikes, error }) => {
-    this.setState({ results: bikes || [], loading: false });
-    this.toggleHeader({ isLoading: false, resultsCount: this.state.results.length });
-    if (error) { this.handleError(error) }
-  }
-
-  handleError = error => {
-    honeybadger.notify(error, { component: "PartialSerialSearch" });
-  }
-
-  toggleHeader = ({ isLoading, resultsCount }) => {
-    const header = document.getElementById("js-partial-serial-search-header");
-    if (!header) { return; }
-
-    header.childNodes.forEach(node => node.classList && node.classList.add("d-none"));
-
-    const titleDisplay = (isLoading)
-          ? "loading"
-          : (resultsCount)
-          ? "loaded"
-          : "loaded-none";
-
-    const sectionTitle = header.getElementsByClassName(titleDisplay)[0]
-    sectionTitle.classList.remove("d-none");
-  }
-
-  render() {
-    if (this.state.loading) {
-      return <Loading />;
-    }
-
-    return (
-      <Fragment>
-        <ul className="bike-boxes">
-          {this.state.results.map(bike => <BikeSearchResult key={bike.id} bike={bike} />)}
-        </ul>
-      </Fragment>
-    );
-  }
+class PartialSerialSearch extends BikeSearch {
+  errorComponentName = "PartialSerialSearch";
+  fetchBikes = fetchPartialMatchSearch;
+  headerDomId = "js-partial-serial-search-header";
 };
 
 export default PartialSerialSearch;
